+++ include/WAVEFORM.h
@@ -32,7 +32,7 @@ struct waveform_ctx{
 };
 
 bool is_waveform(void * waveform_name_index_map, const char *name);
-int populate_rt_offset_arr(void *waveform_name_index_map, uint8_t * rt_variable_offsets, int rt_var_size, void *rt_addresses, struct waveform_ctx *waveforms);
+int populate_rt_offset_arr(void *waveform_name_index_map, uint32_t *rt_variable_offsets, int rt_var_size, void *rt_addresses, struct waveform_ctx *waveforms);
 
 
 #ifdef __cplusplus
